Is the backend enforcing a file size limit?
Resolves #117

Added the ability to set a file size limit for admin and regular users. Enforcing the limit in the DataSetViewSet.upload function.
This will prevent the admin or user to upload file sizes larger than the specified limits while allowing the django website admin to upload files of arbitrarily large sizes.

This fix also introduces a reorganization of the archive api settings into a dictionary.