Various login issues
First login problem,  Resolves #149 - The resolution is to know the workflow.  For a user, who is not a super user, to be activated they must have a person record and either have a role assigned to the user or if the user does not exist in the database yet a user_role assigned to the accompanying user record.  There is a fix in here that makes sure that if the user record already exists and is assigned a role and has a person record then they should be automatically activated.

Message shown when user is authenticated but not activated, Resolves #88 - the issue here was that the message should have been shown on the login page not the home page.

site title is 'Django REST framework', Resolves #136